Create 1_02_class_and_pass_keyword.py
A class is a template for a data type. It describes the kinds of information that class will hold and how a programmer will interact with that data. Its recommended capitalizing the names of classes to make them easier to identify. 

The pass keyword in Python to indicate that the body of the class was intentionally left blank so we don’t cause an IndentationError.