Type() on object, returns class  name = __main__
A class instance is also called an _object_.  
Instantiation takes a class and turns it into an object, the type() function does the opposite of that. 
When called with an object, it returns the class that the object is an instance of.

In Python, __main__ is an identifier used to reference the current file context. When a module is read from standard input, a script, or from an interactive prompt, its __name__ is set equal to __main__.