__init__ method (Constructor)
There are several methods that we can define in a Python class that have special behavior = dunder methods

The __init__() method (note the two underscores before and after the word “init”) is a dunder method is used to initialize a newly created object. 
It is called every time the class is instantiated.

Methods that are used to prepare an object being instantiated are called constructors. 
The word “constructor” is used to describe similar features in other object-oriented programming languages but programmers who refer to a constructor in Python are usually talking about the __init__() method.