Finds the median on a sorted list
The median is the middle number in a sorted sequence of numbers.  
Finding the median of [7, 12, 3, 1, 6] would first consist of sorting the sequence into [1, 3, 6, 7, 12] 
and then locating the middle value, which would be 6.
If you are given a sequence with an even number of elements, you must average the two elements surrounding the middle.