Using a mask to check if the fourth bit is on.
function _check_bit4_ takes one argument _input_, an integer. It should check to see if the fourth bit from the right is on.

A bit mask is just a variable that aids you with bitwise operations. 
A bit mask can help you turn specific bits on, turn others off, 
or just collect data from an integer about which bits are on or off.